init microfluidics device: Do not vent.
To use the new liquid-buffer flow method, after washing the various
lines, we will no longer vent the lines (at least for now), rather,
we'll leave the wash reagent in the channels.

The wash reagent acts as our fluid buffer, and helps avoid (or at the
very least, reduce) the problems caused by air bubbles in the
channels.